add tests for board view and bugs fixed

+++ goals/serializers.py
@@ -39,7 +39,7 @@ class BoardCreateSerializer(serializers.ModelSerializer):
         fields = "__all__"
 
     def create(self, validated_data):
-        user = validated_data.pop("user")
+        user = validated_data.pop("user", self.context["request"].user)
         board = Board.objects.create(**validated_data)
         BoardParticipant.objects.create(
             user=user, board=board, role=BoardParticipant.Role.owner
@@ -57,33 +57,38 @@ class BoardSerializer(serializers.ModelSerializer):
         read_only_fields = ("id", "created", "updated")
 
     def update(self, instance, validated_data):
-        owner = validated_data.pop("user")
-        new_participants = validated_data.pop("participants")
-        new_by_id = {part["user"].id: part for part in new_participants}
+        owner = validated_data.pop("user", self.context["request"].user)
 
-        old_participants = instance.participants.exclude(user=owner)
         with transaction.atomic():
-            for old_participant in old_participants:
-                if old_participant.user_id not in new_by_id:
-                    old_participant.delete()
-                    continue
-                if (
-                    old_participant.role
-                    != new_by_id[old_participant.user_id]["role"]
-                ):
-                    old_participant.role = new_by_id[old_participant.user_id][
-                        "role"
-                    ]
-                    old_participant.save()
-                new_by_id.pop(old_participant.user_id)
-            for new_part in new_by_id.values():
-                BoardParticipant.objects.create(
-                    board=instance,
-                    user=new_part["user"],
-                    role=new_part["role"],
-                )
+            if validated_data.get("participants"):
+                new_participants = validated_data.pop("participants")
+                new_by_id = {
+                    part["user"].id: part for part in new_participants
+                }
 
-            instance.title = validated_data["title"]
+                old_participants = instance.participants.exclude(user=owner)
+
+                for old_participant in old_participants:
+                    if old_participant.user_id not in new_by_id:
+                        old_participant.delete()
+                        continue
+                    if (
+                        old_participant.role
+                        != new_by_id[old_participant.user_id]["role"]
+                    ):
+                        old_participant.role = new_by_id[
+                            old_participant.user_id
+                        ]["role"]
+                        old_participant.save()
+                    new_by_id.pop(old_participant.user_id)
+                for new_part in new_by_id.values():
+                    BoardParticipant.objects.create(
+                        board=instance,
+                        user=new_part["user"],
+                        role=new_part["role"],
+                    )
+            if validated_data.get("title"):
+                instance.title = validated_data.get("title")
             instance.save()
 
         return instance
